pyproject.toml
[project]
name = "aimailcollector"
version = "0.3.4"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24.4",
    "openai>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
    "streamlit>=1.40.1",
    "tqdm>=4.67.1",
]
